backend/requirements.txt
flask
flask-cors
pymongo[srv]
python-dotenv
neo4j
